CalcMe App
CalcMe App for IOS, based on MVC principle. The main code is in
CalcMeViewController. xib file is used only as a interface builder. As
soon as changes are done in xib file they are automatically updated in
CalcMeViewController.designer.cs  Main.cs file is only an entry to the
program.